backend/pyproject.toml
[project]
name = "homepet-backend"
version = "0.1.0"
description = "HomePet Delivery — API Backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "motor>=3.6",
    "beanie>=1.27",
    "pydantic-settings>=2.6",
    "pydantic[email]>=2.0",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "slowapi>=0.1.9",
    "secure>=0.3",
    "python-multipart>=0.0.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.24",
    "httpx>=0.28",
    "ruff>=0.8",
    "mypy>=1.13",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "S", "UP"]

[tool.mypy]
python_version = "3.12"
strict = false
ignore_missing_imports = true
